Begin fleshing out provision NewThriftType model.
Also, touch up EnumMemberElement so it doesn't have an Integer value.
It was always non-null by the time parsing was done, so no point to it.

+++ thrifty-schema/src/main/java/com/microsoft/thrifty/schema/Constant.java
@@ -307,7 +307,7 @@ public class Constant extends Named {
                 if (value.kind() == ConstValueElement.Kind.INTEGER) {
                     long id = (Long) value.value();
                     for (EnumType.Member member : et.members()) {
-                        if (member.value().longValue() == id) {
+                        if (member.value() == id) {
                             return;
                         }
                     }
